pyproject.toml
[project]
name = "projectdash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dylan Chidambaram", email = "dylanchidambaram@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "textual>=8.0.0",
]

[project.scripts]
projectdash = "projectdash.app:run"
pd = "projectdash.cli:main"

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
